Move feeding dependencies to FeedingModule
