+++ shared/feature/feeding/src/debug/java/com/epmedu/animeal/feeding/di/FeedingModule.kt
@@ -5,6 +5,12 @@ import com.epmedu.animeal.api.feeding.FeedingApi
 import com.epmedu.animeal.api.feeding.FeedingPointApi
 import com.epmedu.animeal.api.storage.StorageApi
 import com.epmedu.animeal.auth.AuthAPI
+import com.epmedu.animeal.common.domain.usecase.ForcedArgumentsUseCase
+import com.epmedu.animeal.common.presentation.viewmodel.HomeViewModelEvent
+import com.epmedu.animeal.common.presentation.viewmodel.delegate.ActionDelegate
+import com.epmedu.animeal.common.presentation.viewmodel.delegate.EventDelegate
+import com.epmedu.animeal.common.presentation.viewmodel.delegate.StateDelegate
+import com.epmedu.animeal.common.presentation.viewmodel.handler.error.ErrorHandler
 import com.epmedu.animeal.debugmenu.domain.DebugMenuRepository
 import com.epmedu.animeal.feeding.data.repository.FavouriteRepositoryImpl
 import com.epmedu.animeal.feeding.data.repository.FavouriteRepositoryMock
@@ -16,10 +22,25 @@ import com.epmedu.animeal.feeding.domain.repository.FavouriteRepository
 import com.epmedu.animeal.feeding.domain.repository.FeedingPointRepository
 import com.epmedu.animeal.feeding.domain.repository.FeedingRepository
 import com.epmedu.animeal.feeding.domain.usecase.AddFeedingPointToFavouritesUseCase
+import com.epmedu.animeal.feeding.domain.usecase.CancelFeedingUseCase
+import com.epmedu.animeal.feeding.domain.usecase.FetchCurrentFeedingPointUseCase
+import com.epmedu.animeal.feeding.domain.usecase.FinishFeedingUseCase
+import com.epmedu.animeal.feeding.domain.usecase.GetAllFeedingPointsUseCase
+import com.epmedu.animeal.feeding.domain.usecase.RejectFeedingUseCase
 import com.epmedu.animeal.feeding.domain.usecase.RemoveFeedingPointFromFavouritesUseCase
+import com.epmedu.animeal.feeding.domain.usecase.StartFeedingUseCase
+import com.epmedu.animeal.feeding.domain.usecase.UpdateAnimalTypeSettingsUseCase
+import com.epmedu.animeal.feeding.presentation.viewmodel.FeedingPointState
+import com.epmedu.animeal.feeding.presentation.viewmodel.handler.feeding.DefaultFeedingHandler
+import com.epmedu.animeal.feeding.presentation.viewmodel.handler.feeding.FeedingHandler
+import com.epmedu.animeal.feeding.presentation.viewmodel.handler.feedingpoint.DefaultFeedingPointHandler
+import com.epmedu.animeal.feeding.presentation.viewmodel.handler.feedingpoint.FeedingPointHandler
+import com.epmedu.animeal.router.presentation.RouteHandler
+import com.epmedu.animeal.timer.presentation.handler.TimerHandler
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
+import dagger.hilt.android.scopes.ViewModelScoped
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
 import kotlinx.coroutines.Dispatchers
@@ -109,4 +130,56 @@ object FeedingModule {
     fun providesDeleteFavouriteFeedingPointUseCase(
         repo: FavouriteRepository
     ) = RemoveFeedingPointFromFavouritesUseCase(repo)
+
+    @ViewModelScoped
+    @Provides
+    fun providesFeedingHandler(
+        stateDelegate: StateDelegate<FeedingPointState>,
+        actionDelegate: ActionDelegate,
+        routeHandler: RouteHandler,
+        errorHandler: ErrorHandler,
+        feedingPointHandler: FeedingPointHandler,
+        timerHandler: TimerHandler,
+        fetchCurrentFeedingPointUseCase: FetchCurrentFeedingPointUseCase,
+        startFeedingUseCase: StartFeedingUseCase,
+        cancelFeedingUseCase: CancelFeedingUseCase,
+        rejectFeedingUseCase: RejectFeedingUseCase,
+        finishFeedingUseCase: FinishFeedingUseCase,
+        forcedArgumentsUseCase: ForcedArgumentsUseCase
+    ): FeedingHandler = DefaultFeedingHandler(
+        stateDelegate,
+        actionDelegate,
+        routeHandler,
+        errorHandler,
+        feedingPointHandler,
+        timerHandler,
+        fetchCurrentFeedingPointUseCase,
+        startFeedingUseCase,
+        cancelFeedingUseCase,
+        rejectFeedingUseCase,
+        finishFeedingUseCase,
+        forcedArgumentsUseCase
+    )
+
+    @ViewModelScoped
+    @Provides
+    fun providesFeedingPointHandler(
+        stateDelegate: StateDelegate<FeedingPointState>,
+        eventDelegate: EventDelegate<HomeViewModelEvent>,
+        actionDelegate: ActionDelegate,
+        errorHandler: ErrorHandler,
+        getAllFeedingPointsUseCase: GetAllFeedingPointsUseCase,
+        addFeedingPointToFavouritesUseCase: AddFeedingPointToFavouritesUseCase,
+        removeFeedingPointFromFavouritesUseCase: RemoveFeedingPointFromFavouritesUseCase,
+        updateAnimalTypeSettingsUseCase: UpdateAnimalTypeSettingsUseCase,
+    ): FeedingPointHandler = DefaultFeedingPointHandler(
+        stateDelegate,
+        eventDelegate,
+        actionDelegate,
+        errorHandler,
+        getAllFeedingPointsUseCase,
+        addFeedingPointToFavouritesUseCase,
+        removeFeedingPointFromFavouritesUseCase,
+        updateAnimalTypeSettingsUseCase
+    )
 }
